fix extended eslint config
