+++ packages/eslint-config/next.js
@@ -1,11 +1,7 @@
-/**
- * @template {import('@types/eslint').ESLint.ConfigData} T
- * @param {T} config A generic parameter that flows through to the return type
- * @constraint {{import('@types/eslint').ESLint.ConfigData}}
- */
-
+/* eslint-disable import/no-unused-modules */
+/** @type {import('eslint').ESLint.ConfigData} */
 const config = {
-	extends: ['../config', 'next/core-web-vitals'],
+	extends: ['./', 'next/core-web-vitals'],
 	rules: {
 		'@next/next/no-html-link-for-pages': 'off',
 	},
