+++ packages/eslint-config/storybook.js
@@ -1,0 +1,8 @@
+/* eslint-disable import/no-unused-modules */
+/** @type {import('eslint').ESLint.ConfigData} */
+const config = {
+	plugins: ['storybook'],
+	extends: ['./', 'plugin:storybook/recommended'],
+}
+
+module.exports = config
